+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 -2.0.0
     - Change Token Hashing Vectorizer hashing function
     - Gradient Boost always uses mean base estimator
+    - Move dummy estimators to the Extras package
 
 - 1.2.0
     - Added Logit Boost classifier